pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "macinput-mcp"
version = "0.1.0"
description = "macOS keyboard, mouse, and screenshot automation exposed as an MCP server."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "sigma711" }
]
keywords = ["macos", "automation", "mcp", "mouse", "keyboard", "screenshot"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: MacOS X",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities"
]
dependencies = [
  "mcp[cli]>=1.0.0,<2.0.0",
  "pyobjc-framework-Quartz>=10.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<0.12"
]

[project.scripts]
macinput-mcp = "macinput.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/macinput"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
